Make and use a helper for some of the duplication
